Use spaces instead of tabs everywhere
Let's just get this over with while it's still possible. I'm tired
of editors and code previews defaulting to a tab width of 8, when
I work with four spaces in Nicotine+ and other projects.

Remove some redundant comments in the process. For server messages,
check the Nicotine+ protocol documentation instead.