Use epoll to watch sockets on Linux
This will allow for increased performance when watching a lot of
sockets. Operating systems other than Linux will fall back to
select().

A kqueue selector for macOS and the various BSDs will be added in
the near future.

The API is close to Python's selectors module, which we use in
Nicotine+.